requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.6.0
python-dotenv>=1.0.1
sentence-transformers>=2.3.1
faiss-cpu>=1.7.4
requests>=2.31.0
google-genai>=0.3.0
openai>=1.10.0
numpy
tiktoken
# Do testów
pytest
httpx

